requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
aiosqlite
streamlit
pandas
matplotlib
plotly